requirements.txt
httpx
beautifulsoup4
trafilatura
pandas==2.2.3
openpyxl==3.1.5
anthropic
python-dotenv==1.0.1
playwright==1.49.1
celery
redis
psycopg2-binary
SQLAlchemy
flower
